Components Styles and Transparent Status Bar
Changed various component styles and made the status bar transparent. Moved Anchor Text to "texts" folder
