--- src/myapp/components/anchor-text.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import {Text} from '@ui-kitten/components';
-import style from '../styles/style';
-
-const AnchorText = (props): React.ReactElement => {
-  return (
-    <Text style={[style.anchorText, props.style]} onPress={props.onPress}>
-      {props.children}
-    </Text>
-  );
-};
-
-export default AnchorText;
