+++ src/myapp/components/buttons/close-button.tsx
@@ -1,19 +1,30 @@
 import React from 'react';
-import style from '../../styles/style';
+import {StyleSheet} from 'react-native';
 import {Icon} from '@ui-kitten/components';
+// Global styles.
 import globalColors from '../../styles/colors';
 
-const CloseButton = ({navigation}): React.ReactElement => {
+const CloseButton = (props): React.ReactElement => {
   return (
     <Icon
-      style={style.closeButton}
-      height={40}
-      onPress={navigation.goBack}
-      width={40}
+      style={[styles.closeButton, props.style]}
+      height={35}
+      onPress={props.navigation.goBack}
+      width={35}
       fill={globalColors.greenSecondary}
       name="close-outline"
     />
   );
 };
+
+const styles = StyleSheet.create({
+  closeButton: {
+    width: 40,
+    height: 40,
+    marginLeft: -5,
+    marginTop: 15,
+    marginBottom: 10,
+  },
+});
 
 export default CloseButton;
