+++ src/myapp/layouts/start/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {View, Image, StyleSheet} from 'react-native';
 import {Text} from '@ui-kitten/components';
 // My components
-import AnchorText from '../../components/anchor-text';
+import AnchorText from '../../components/texts/anchor-text';
 import CustomButton from '../../components/buttons/custom-button';
 import DefaultLayout from '../../components/default-layout';
 // Global styles.
