+++ src/myapp/styles/colors.ts
@@ -4,9 +4,9 @@ const globalColors = {
   lightGreen: '#cfe2d5',
 
   white: '#fff',
-  darkGray: '#707070',
-  darkerGray: '#373737',
-  lightGray: '#a7a7a7',
+  darkGray: '#707070', // rgb(112, 112, 112)
+  darkerGray: '#373737', // rgb(55, 55, 55)
+  lightGray: '#a7a7a7', // rgb(167, 167, 167)
   black: '#000',
 
   backgroundDefault: '#e6f0e9',
