+++ src/myapp/styles/vars.ts
@@ -1,6 +1,8 @@
 const globalVars = {
   outsidePadding: 24,
   regularSpacing: 24,
+  fontBold: 'Montserrat-Bold',
+  fontRegular: 'Montserrat-Medium',
 };
 
 export default globalVars;
